% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toy.R
\name{get_toy_gene_list}
\alias{get_toy_gene_list}
\title{Get toy gene list filename}
\usage{
get_toy_gene_list()
}
\value{
The name of a toy gene list file
}
\description{
Get toy gene list filename
}
\examples{
toy_gene_list <- get_toy_gene_list()
}

