% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toy.R
\name{get_toy_rpkm}
\alias{get_toy_rpkm}
\title{Get toy rpkm filename}
\usage{
get_toy_rpkm()
}
\value{
The name of a toy rpkm file
}
\description{
Get toy rpkm filename
}
\examples{
toy_rpkm <- get_toy_rpkm()
}

