% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmapHIV.R
\name{heatmapHIV}
\alias{heatmapHIV}
\title{Produce a heatmap}
\usage{
heatmapHIV(counts, gene_list)
}
\arguments{
\item{counts}{The filename of the RPKM or raw count matrix.}

\item{gene_list}{The filename of the gene list to subset the matrix before
producing the heatmap.}
}
\value{
Invisibly return the \code{matrix} used to produce the heatmap.
}
\description{
Produce a heatmap
}
\examples{
\dontrun{
heatmapHIV("RPKM.txt", "gene_list.txt")
}
}

